REF: share join methods for DTI/TDI
